backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.2.2
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-timezone-field==7.1
djangorestframework==3.16.0
filelock==3.18.0
fsspec==2025.5.1
huggingface-hub==0.32.4
idna==3.10
imaplib2==3.6
kombu==5.5.4
numpy==2.2.6
packaging==25.0
prompt_toolkit==3.0.51
python-crontab==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==6.2.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
six==1.17.0
sqlparse==0.5.3
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.52.4
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
